Deprecating api/utils.py module
