+++ robottelo/host_helpers/api_factory.py
@@ -2,12 +2,14 @@
 It is not meant to be used directly, but as part of a robottelo.hosts.Satellite instance
 example: my_satellite.api_factory.api_method()
 """
+import time
 from contextlib import contextmanager
 
 from fauxfactory import gen_ipaddr
 from fauxfactory import gen_mac
 from fauxfactory import gen_string
 from nailgun import entity_mixins
+from nailgun.client import request
 from nailgun.entity_mixins import call_entity_method_with_timeout
 from requests import HTTPError
 
@@ -630,3 +632,129 @@ class APIFactory:
         if not template.locked:
             template.locked = True
             template.update(['locked'])
+
+    def attach_custom_product_subscription(self, prod_name=None, host_name=None):
+        """Attach custom product subscription to client host
+        :param str prod_name: custom product name
+        :param str host_name: client host name
+        """
+        host = self._satellite.api.Host().search(query={'search': f'{host_name}'})[0]
+        product_subscription = self._satellite.api.Subscription().search(
+            query={'search': f'name={prod_name}'}
+        )[0]
+        self._satellite.api.HostSubscription(host=host.id).add_subscriptions(
+            data={'subscriptions': [{'id': product_subscription.id, 'quantity': 1}]}
+        )
+
+    def wait_for_errata_applicability_task(
+        self, host_id, from_when, search_rate=1, max_tries=10, poll_rate=None, poll_timeout=15
+    ):
+        """Search the generate applicability task for given host and make sure it finishes
+
+        :param int host_id: Content host ID of the host where we are regenerating applicability.
+        :param int from_when: Timestamp (in UTC) to limit number of returned tasks to investigate.
+        :param int search_rate: Delay between searches.
+        :param int max_tries: How many times search should be executed.
+        :param int poll_rate: Delay between the end of one task check-up and
+                the start of the next check-up. Parameter for
+                ``nailgun.entities.ForemanTask.poll()`` method.
+        :param int poll_timeout: Maximum number of seconds to wait until timing out.
+                Parameter for ``nailgun.entities.ForemanTask.poll()`` method.
+        :return: Relevant errata applicability task.
+        :raises: ``AssertionError``. If not tasks were found for given host until timeout.
+        """
+        assert isinstance(host_id, int), 'Param host_id have to be int'
+        assert isinstance(from_when, int), 'Param from_when have to be int'
+        now = int(time.time())
+        assert from_when <= now, 'Param from_when have to be timestamp in the past'
+        for _ in range(max_tries):
+            now = int(time.time())
+            max_age = now - from_when + 1
+            search_query = (
+                '( label = Actions::Katello::Host::GenerateApplicability OR label = '
+                'Actions::Katello::Host::UploadPackageProfile ) AND started_at > "%s seconds ago"'
+                % max_age
+            )
+            tasks = self._satellite.api.ForemanTask().search(query={'search': search_query})
+            tasks_finished = 0
+            for task in tasks:
+                if (
+                    task.label == 'Actions::Katello::Host::GenerateApplicability'
+                    and host_id in task.input['host_ids']
+                ):
+                    task.poll(poll_rate=poll_rate, timeout=poll_timeout)
+                    tasks_finished += 1
+                elif (
+                    task.label == 'Actions::Katello::Host::UploadPackageProfile'
+                    and host_id == task.input['host']['id']
+                ):
+                    task.poll(poll_rate=poll_rate, timeout=poll_timeout)
+                    tasks_finished += 1
+            if tasks_finished > 0:
+                break
+            time.sleep(search_rate)
+        else:
+            raise AssertionError(
+                f"No task was found using query '{search_query}' for host '{host_id}'"
+            )
+
+    def wait_for_syncplan_tasks(self, repo_backend_id=None, timeout=10, repo_name=None):
+        """Search the pulp tasks and identify repositories sync tasks with
+        specified name or backend_identifier
+
+        :param repo_backend_id: The Backend ID for the repository to identify the
+            repo in Pulp environment
+        :param timeout: Value to decided how long to check for the Sync task
+        :param repo_name: If repo_backend_id can not be passed, pass the repo_name
+        """
+        if repo_name:
+            repo_backend_id = (
+                self._satellite.api.Repository()
+                .search(query={'search': f'name="{repo_name}"', 'per_page': '1000'})[0]
+                .backend_identifier
+            )
+        # Fetch the Pulp password
+        pulp_pass = self._satellite.execute(
+            'grep "^default_password" /etc/pulp/server.conf | awk \'{print $2}\''
+        ).stdout.splitlines()[0]
+        # Set the Timeout value
+        timeup = time.time() + int(timeout) * 60
+        # Search Filter to filter out the task based on backend-id and sync action
+        filtered_req = {
+            'criteria': {
+                'filters': {
+                    'tags': {'$in': [f"pulp:repository:{repo_backend_id}"]},
+                    'task_type': {'$in': ["pulp.server.managers.repo.sync.sync"]},
+                }
+            }
+        }
+        while True:
+            if time.time() > timeup:
+                raise self._satellite.api.APIResponseError(
+                    f'Pulp task with repo_id {repo_backend_id} not found'
+                )
+            # Send request to pulp API to get the task info
+            req = request(
+                'POST',
+                f'{self._satellite.url}/pulp/api/v2/tasks/search/',
+                verify=False,
+                auth=('admin', f'{pulp_pass}'),
+                headers={'content-type': 'application/json'},
+                data=filtered_req,
+            )
+            # Check Status code of response
+            if req.status_code != 200:
+                raise self._satellite.api.APIResponseError(
+                    f'Pulp task with repo_id {repo_backend_id} not found'
+                )
+            # Check content of response
+            # It is '[]' string for empty content when backend_identifier is wrong
+            if len(req.content) > 2:
+                if req.json()[0].get('state') in ['finished']:
+                    return True
+                elif req.json()[0].get('error'):
+                    raise AssertionError(
+                        f"Pulp task with repo_id {repo_backend_id} error or not found: "
+                        f"'{req.json().get('error')}'"
+                    )
+            time.sleep(2)
