--- tests/robottelo/test_api_utils.py
@@ -1,1 +1,0 @@
-"""Unit tests for :mod:`robottelo.api.utils`."""
